chore: prepare Release 24 for release testing
